fix(graph): align stage table rows with height
